requirements.txt
matplotlib==3.9.2
pandas==2.2.2
numpy==2.0.1
prophet==1.1.5
python-dotenv==1.0.1
ShopifyAPI==12.6.0